+++ CHANGELOG.md
@@ -17,6 +17,7 @@ https://github.com/gab1618/lapse/compare/latest..HEAD
 
 - Send: show only body on request with the flag `--body`
 - Send: dry run requests to see preview using `--dry-run`
+- Ls: show method and url for endpoints
 
 ## [0.4.0] - 2026-08-28